chore: refactor doctor create mealplan
